api/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
sqlmodel==0.0.8
pg8000==1.30.3
httpx==0.24.1
pyjwt==2.8.0
python-multipart==0.0.6
email-validator==2.0.0
